admin/app/requirements/base.txt
django==3.2
python-dotenv==0.21.0
psycopg2==2.9.4
django_split_settings==1.2.0
uwsgi==2.0.20
dj-static==0.0.6
pydantic[email]==1.10.2
orjson==3.8.0
aio_pika==8.2.4
backoff==2.2.1
